+++ examples/readme_example.py
@@ -1,13 +1,15 @@
 from src.population import create_bacteria
-from src.research.statistic import Stats, ResearchParams
+from src.research.statistic import PopulationResearch, ResearchParams
 
-init_group = [create_bacteria(5, 0.6, 0.2) for i in range(5)]
 
-stats = Stats(init_group)
+research = PopulationResearch()
+research.add_individuals([create_bacteria(5, 0.6, 0.2) for _ in range(10)])
 
-iteration_params = ResearchParams('uniform', 'default', 'normal', 'default')
+print(research)
+
+iteration_params = ResearchParams('uniform', 1, 'normal', 0.0001)  # тут scale для мутатора и селектора и их тип
 
 # Получить возможные строки(можно пихать эти массивы в выпадающие окошки пока что):
 print(ResearchParams.get_modes())
 
-print(stats.research(10, iteration_params).data) # -> фрейм с 10-ю строками, можно сделать не 10, а 1 и на каждой итерации менять params
+print(research.research(4, iteration_params).data) # -> фрейм с 10-ю строками, можно сделать не 10, а 1 и на каждой итерации менять params
