+++ src/main.py
@@ -1,11 +1,7 @@
-import dash_html_components as html
-import dash_core_components as dcc
 from ui.ui import DashUI
 from src.ui.pages.Home.home_page import HomePage
 from callback_logic import *
 
-def t():
-    return [-100]
 
 class App:
     def __init__(self):
@@ -14,6 +10,7 @@ class App:
 
     def run(self):
         self.ui.run()
+
 
 
 if __name__ == "__main__":
