+++ src/population/__init__.py
@@ -1,5 +1,4 @@
 from .mutations.normal_mutator import *
-from .life_cycle import LifeCycle
-from src.population.individuals.bacteria import *
-from .population import *
-from .selectors.bacterial_selector import *
+from src.population.species.bacteria.bacteria import *
+from src.population.populations.simple_population import *
+from .selectors.genomic_selector import *
