+++ src/research/statistic.py
@@ -1,55 +1,13 @@
-from copy import deepcopy
-from dataclasses import dataclass, field
-from typing import ClassVar, Dict
-from src.population.mutations.mutator_parameters import MutatorParams
-from src.population.population import Population
-from src.population import AbstractSelector, NormalMutator, SelectorParams, UniformSelector
-from src.population.life_cycle import LifeCycle
+from typing import Dict
+from src.population.populations.simple_population import Population
+from src.population import AbstractSelector, AbstractMutator
 import pandas as pd
+from dataclasses import dataclass, field
+from src.research.research_params import ResearchRes, ResearchParams
 
 
-@dataclass
-class ResearchParams:
-    selector: str
-    selector_mode: str
-    mutator: str
-    mutator_mode: str
-    _selector_types: ClassVar = field(init=False, default={'uniform': UniformSelector})
-    _selector_modes: ClassVar = field(init=False, default={'default': 1, 'cruel': 1.3, 'loyal': 0.5})
-    _mutator_types: ClassVar = field(init=False, default={'normal': NormalMutator})
-    _mutator_modes: ClassVar = field(init=False, default={'default': 0.01, 'high': 0.1, 'low': 0.005})
-
-    def convert(self):
-        selector_params = SelectorParams(0, ResearchParams._selector_modes[self.selector_mode])
-        mutator_params = MutatorParams(0, ResearchParams._mutator_modes[self.mutator_mode])
-
-        return (
-            ResearchParams._selector_types[self.selector](selector_params),
-            ResearchParams._mutator_types[self.mutator](mutator_params)
-        )
-
-    @staticmethod
-    def get_modes():
-        return {'Selector types': list(ResearchParams._selector_types.keys()),
-                'Selector modes': list(ResearchParams._selector_modes.keys()),
-                'Mutator types': list(ResearchParams._mutator_types.keys()),
-                'Mutator modes': list(ResearchParams._mutator_modes.keys())}
-
-@dataclass
-class ResearchRes:
-    id: int
-    data: pd.DataFrame
-    params: ResearchParams
-
-
-def _get_stats(population: Population) -> Dict[str, int]:
-    all_n, alive_n = len(population.get_all()), len(population.get_alive())
-    dead_n = all_n - alive_n
-    print({'all': all_n, 'alive': alive_n, 'dead': dead_n})
-    return {'all': all_n, 'alive': alive_n, 'dead': dead_n}
-
-
-class Stats:
+@dataclass(frozen=True)
+class PopulationResearch:
     """
         Class with some statistical tools for population analysis.
 
@@ -62,11 +20,12 @@ class Stats:
         -------
         num_of_individuals(self, num_iter: int, selectors: AbstractSelector, mutator: AbstractMutator,
                            draw_func) -> DataFrame
-        Show number of individuals in population
+        Show number of species in population
     """
+    _population: Population = field(default_factory=Population)
 
-    def __init__(self, bacterias: list):
-        self._bacterias = bacterias
+    def add_individuals(self, individuals: list):
+        self._population.add(individuals)
 
     def research(self,
                  num_iter: int,
@@ -90,13 +49,23 @@ class Stats:
         DataFrame
             Table with state of population on each iteration
         """
-        population = Population(deepcopy(self._bacterias))
-        cycle = LifeCycle(population)
-        iter_params = params.convert()
-        fr = pd.DataFrame(columns=['all', 'alive', 'dead'])
+        fr = Stats.get_empty_frame()
 
         for _ in range(num_iter):
-            cycle.iterate(*iter_params)
-            fr = fr.append(_get_stats(population), ignore_index=True)
+            self._population.iterate(*params.iter_params())
+            fr = fr.append(Stats.get_stats(self._population), ignore_index=True)
 
         return ResearchRes(0, fr, params)
+
+
+class Stats:
+    @staticmethod
+    def get_empty_frame():
+        return pd.DataFrame(columns=['all', 'alive', 'dead'])
+
+    @staticmethod
+    def get_stats(population: Population) -> Dict[str, int]:
+        all_n, alive_n = len(population.get_all()), len(population.get_alive())
+        dead_n = all_n - alive_n
+
+        return {'all': all_n, 'alive': alive_n, 'dead': dead_n}
